test(riasec): keep rollout gate non-runtime
