+++ docs/codex/pr-train-state.json
@@ -1,6 +1,6 @@
 {
   "train_id": "email-first-result-access",
-  "updated_at": "2026-06-22T12:13:27+08:00",
+  "updated_at": "2026-06-22T12:24:31+08:00",
   "FREEMIUM-LOCALE-POLICY-SCAN-01": {
     "repo": "fap-api",
     "branch": "codex/freemium-locale-policy-scan-01",
@@ -57676,7 +57676,7 @@
     "repo": "fap-api",
     "branch": "codex/riasec-result-production-rollout-gate-01",
     "base": "main",
-    "status": "pull_request_open",
+    "status": "github_checks_failed_scoped_fix_local_passed",
     "train_scope": "riasec_result_ops_agent_release_chain",
     "title": "RIASEC-RESULT-PRODUCTION-ROLLOUT-GATE-01: test(riasec): add production rollout gate",
     "depends_on": [
@@ -57690,7 +57690,17 @@
       "git_diff_check": "passed",
       "scope_validation": "passed",
       "branch_pushed": "passed",
-      "pr_created": "passed"
+      "pr_created": "passed",
+      "github_verify_bigfive_initial": "failed: BigFiveResultPageV2CoreBodyPreviewTest flagged newly added backend/app/Services/Riasec/Rollout production rollout service files as runtime path diffs.",
+      "github_failure_inspected": "passed",
+      "github_failure_fix_scope": "passed: removed new runtime service files and retained gate behavior in allowed test fixture/governance/config scope.",
+      "php_artisan_test_filter_RiasecResultPage_after_github_fix": "passed",
+      "php_artisan_test_filter_ProductionRuntime_after_github_fix": "passed",
+      "php_artisan_test_filter_RolloutGate_after_github_fix": "passed",
+      "php_l_RolloutGateTest_after_github_fix": "passed",
+      "git_diff_check_after_github_fix": "passed",
+      "scope_validation_after_github_fix": "passed",
+      "php_artisan_test_bigfive_runtime_freeze_guard_after_github_fix": "passed"
     },
     "failure_reason": null,
     "commit_sha": "20a7cf5d649c076557f8ffc77e3f455e65b4fa41",
@@ -57725,6 +57735,18 @@
         "from": "local_checks_passed",
         "to": "pull_request_open",
         "notes": "Opened GitHub PR #2276 after local validation and scope validation passed. Production rollout remains manual-only and not executed."
+      },
+      {
+        "at": "2026-06-22T12:24:31+08:00",
+        "from": "pull_request_open",
+        "to": "github_checks_failed_fix_in_scope",
+        "notes": "Inspected PR #2276 verify-bigfive failure. The current-scope repair removes new RIASEC Rollout service files that triggered BigFive runtime freeze guard and moves gate simulation to RolloutGateTest fixture; no production rollout automation is introduced."
+      },
+      {
+        "at": "2026-06-22T12:24:31+08:00",
+        "from": "github_checks_failed_fix_in_scope",
+        "to": "github_checks_failed_scoped_fix_local_passed",
+        "notes": "Local validation passed after scoped GitHub fix: RiasecResultPage, ProductionRuntime, RolloutGate, php -l, git diff --check, BigFive runtime freeze guard, and scope validation."
       }
     ]
   },
